Improve templates
The idea behind adding the "link down" option was to force
the guest's NetworkManger to the management interface as the main
interface. However, this is making vhost-user interfaces ususable
(they cannot be dynamically modified so virsh domif-setlink doesn't
work.

Something else must be done. For now, reverting that change and
improving the PCIe topology